refactor: export 타입 index.ts로 종합
todo: image ts 수정
