--- src/types/CardList.ts
@@ -1,6 +1,0 @@
-import CardType from './Card';
-
-export default interface CardList {
-  cards: CardType[];
-  setCards: React.Dispatch<React.SetStateAction<CardType[]>>;
-}
