--- src/types/CardNumbers.ts
@@ -1,6 +1,0 @@
-export interface CardNumbersType {
-  0: string;
-  1: string;
-  2: string;
-  3: string;
-}
