--- src/types/ExpiredDate.ts
@@ -1,4 +1,0 @@
-export interface ExpiredDatesType {
-  0: string;
-  1: string;
-}
